+++ apps/genetics/src/ot-ui-components/components/Button.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { default as MuiButton } from '@material-ui/core/Button';
+import { ButtonProps, default as MuiButton } from '@material-ui/core/Button';
 
-const Button = ({ children, color, variant, ...rest }) => (
+const Button = ({ children, color, variant, ...rest }: ButtonProps) => (
   <MuiButton
     color={color ? color : 'primary'}
     variant={variant ? variant : 'contained'}
